context : move interface implementation to source file + factory
ggml-ci
